rename files in input
