+++ qiskit/aqua/input/classification_input.py
